Allow types to wrap other types
This allows types to wrap other types like:

```js
type.convert(type.check(Number))
```

And flip a `isStrict` bit along the way.

This is a large refactor of how can-type works and should be more memory
efficient now, as it no longer creates a bunch of new objects in a
closure for each type and largely uses the proto chain for inheritance.

This uses the `isStrict` method to achieve flipping strictness, but
opens the door to a smaller refactor to remove this need.

This also makes it so that:

```js
type.convert(type.check(String)) === type.convert(type.check(String))
```

However, it's not currently true that:

js
```
type.convert(type.maybe(String)) === type.maybeConvert(String));
```

Closes #27