Refactor
This is a refactor to use a new symbol `strictTypeOf` that points to a
convertible TypeObject. The other methods, check, etc. use this as a
base and then add a strict checking `can.new` (in the case of the strict
methods) and/or a maybe check for `can.isMember`. They all also
keep track of adding maybes to the schema if needed.