Add default types for common needs
This adds the following:

* type.String
* type.Number
* type.Boolean
* type.Date
* type.Integer
* type.ToString
* type.ToNumber
* type.ToBoolean
* type.ToDate
* type.ToInteger

As outlined in https://github.com/canjs/canjs/issues/5309

Omitted from this PR are:

* type.Object
* type.Array
* type.DeepObservable

Because of circular dependencies we can't have `can-type` importing
can-observable-object, can-observable-array, or can-deep-observable.

Because of tree-shaking we also can't have those types *add* the types
to the `type.` namespace. It would only work if the user was already
importing `ObservableObject`, for example. If they are already importing
ObservableObject, they already have a strict type to use.

This leaves open the possibility of a *converting* type on
`ObservableObject`, etc. This brings up the naming issue. To keep in
line with this PR you would expect `ObservableObject.ToObject`. I think
we should first consider an alternative to the [converting type
names](https://github.com/canjs/canjs/issues/5309#issuecomment-539512929),
for example perhaps `ObservableObject.CoercedObject` would be better?